Rename the sparse backend completely to "sparse"
Having it be called "numpy" in some places causes issues.
